requirements.txt
torch-ema
torchmetrics
ninja
trimesh
opencv-python
tensorboardX
numpy==1.24
pandas
tqdm
matplotlib==3.5.2
PyMCubes
rich
pysdf
dearpygui
packaging
scipy
lpips
imageio==2.13.0
torchmetrics
imageio-ffmpeg==0.4.8
open3d
configargparse
scikit-image
nksr
black
pyquaternion
camtools==0.1.3
natsort
ujson
pyglet==1.5.29